# [SPARK-36444][SQL] Remove OptimizeSubqueries from batch of PartitionPruning