--- a/sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q58.sf100/simplified.txt
+++ b/sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q58.sf100/simplified.txt
@@ -18,30 +18,32 @@ TakeOrderedAndProject [item_id,ss_item_rev,ss_dev,cs_item_rev,cs_dev,ws_item_rev
                                   ColumnarToRow
                                     InputAdapter
                                       Scan parquet default.store_sales [ss_item_sk,ss_ext_sales_price,ss_sold_date_sk]
-                                InputAdapter
-                                  BroadcastExchange #2
-                                    WholeStageCodegen (2)
-                                      Project [d_date_sk]
-                                        BroadcastHashJoin [d_date,d_date]
-                                          Filter [d_date_sk]
-                                            ColumnarToRow
-                                              InputAdapter
-                                                Scan parquet default.date_dim [d_date_sk,d_date]
-                                          InputAdapter
-                                            BroadcastExchange #3
-                                              WholeStageCodegen (1)
-                                                Project [d_date]
-                                                  Filter [d_week_seq]
-                                                    Subquery #1
+                                        SubqueryBroadcast [d_date_sk] #1
+                                          BroadcastExchange #2
+                                            WholeStageCodegen (2)
+                                              Project [d_date_sk]
+                                                BroadcastHashJoin [d_date,d_date]
+                                                  Filter [d_date_sk]
+                                                    ColumnarToRow
+                                                      InputAdapter
+                                                        Scan parquet default.date_dim [d_date_sk,d_date]
+                                                  InputAdapter
+                                                    BroadcastExchange #3
                                                       WholeStageCodegen (1)
-                                                        Project [d_week_seq]
-                                                          Filter [d_date]
+                                                        Project [d_date]
+                                                          Filter [d_week_seq]
+                                                            Subquery #2
+                                                              WholeStageCodegen (1)
+                                                                Project [d_week_seq]
+                                                                  Filter [d_date]
+                                                                    ColumnarToRow
+                                                                      InputAdapter
+                                                                        Scan parquet default.date_dim [d_date,d_week_seq]
                                                             ColumnarToRow
                                                               InputAdapter
                                                                 Scan parquet default.date_dim [d_date,d_week_seq]
-                                                    ColumnarToRow
-                                                      InputAdapter
-                                                        Scan parquet default.date_dim [d_date,d_week_seq]
+                                InputAdapter
+                                  ReusedExchange [d_date_sk] #2
                             InputAdapter
                               BroadcastExchange #4
                                 WholeStageCodegen (3)
@@ -66,6 +68,7 @@ TakeOrderedAndProject [item_id,ss_item_rev,ss_dev,cs_item_rev,cs_dev,ws_item_rev
                                         ColumnarToRow
                                           InputAdapter
                                             Scan parquet default.catalog_sales [cs_item_sk,cs_ext_sales_price,cs_sold_date_sk]
+                                              ReusedSubquery [d_date_sk] #1
                                       InputAdapter
                                         ReusedExchange [d_date_sk] #2
                                   InputAdapter
@@ -87,6 +90,7 @@ TakeOrderedAndProject [item_id,ss_item_rev,ss_dev,cs_item_rev,cs_dev,ws_item_rev
                                     ColumnarToRow
                                       InputAdapter
                                         Scan parquet default.web_sales [ws_item_sk,ws_ext_sales_price,ws_sold_date_sk]
+                                          ReusedSubquery [d_date_sk] #1
                                   InputAdapter
                                     ReusedExchange [d_date_sk] #2
                               InputAdapter