--- a/sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q83/simplified.txt
+++ b/sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q83/simplified.txt
@@ -17,38 +17,40 @@ TakeOrderedAndProject [item_id,sr_item_qty,sr_dev,cr_item_qty,cr_dev,wr_item_qty
                                 ColumnarToRow
                                   InputAdapter
                                     Scan parquet default.store_returns [sr_item_sk,sr_return_quantity,sr_returned_date_sk]
+                                      SubqueryBroadcast [d_date_sk] #1
+                                        BroadcastExchange #2
+                                          WholeStageCodegen (3)
+                                            Project [d_date_sk]
+                                              BroadcastHashJoin [d_date,d_date]
+                                                Filter [d_date_sk]
+                                                  ColumnarToRow
+                                                    InputAdapter
+                                                      Scan parquet default.date_dim [d_date_sk,d_date]
+                                                InputAdapter
+                                                  BroadcastExchange #3
+                                                    WholeStageCodegen (2)
+                                                      Project [d_date]
+                                                        BroadcastHashJoin [d_week_seq,d_week_seq]
+                                                          ColumnarToRow
+                                                            InputAdapter
+                                                              Scan parquet default.date_dim [d_date,d_week_seq]
+                                                          InputAdapter
+                                                            BroadcastExchange #4
+                                                              WholeStageCodegen (1)
+                                                                Project [d_week_seq]
+                                                                  Filter [d_date]
+                                                                    ColumnarToRow
+                                                                      InputAdapter
+                                                                        Scan parquet default.date_dim [d_date,d_week_seq]
                               InputAdapter
-                                BroadcastExchange #2
+                                BroadcastExchange #5
                                   WholeStageCodegen (1)
                                     Filter [i_item_sk,i_item_id]
                                       ColumnarToRow
                                         InputAdapter
                                           Scan parquet default.item [i_item_sk,i_item_id]
                           InputAdapter
-                            BroadcastExchange #3
-                              WholeStageCodegen (4)
-                                Project [d_date_sk]
-                                  BroadcastHashJoin [d_date,d_date]
-                                    Filter [d_date_sk]
-                                      ColumnarToRow
-                                        InputAdapter
-                                          Scan parquet default.date_dim [d_date_sk,d_date]
-                                    InputAdapter
-                                      BroadcastExchange #4
-                                        WholeStageCodegen (3)
-                                          Project [d_date]
-                                            BroadcastHashJoin [d_week_seq,d_week_seq]
-                                              ColumnarToRow
-                                                InputAdapter
-                                                  Scan parquet default.date_dim [d_date,d_week_seq]
-                                              InputAdapter
-                                                BroadcastExchange #5
-                                                  WholeStageCodegen (2)
-                                                    Project [d_week_seq]
-                                                      Filter [d_date]
-                                                        ColumnarToRow
-                                                          InputAdapter
-                                                            Scan parquet default.date_dim [d_date,d_week_seq]
+                            ReusedExchange [d_date_sk] #2
             InputAdapter
               BroadcastExchange #6
                 WholeStageCodegen (11)
@@ -65,10 +67,11 @@ TakeOrderedAndProject [item_id,sr_item_qty,sr_dev,cr_item_qty,cr_dev,wr_item_qty
                                       ColumnarToRow
                                         InputAdapter
                                           Scan parquet default.catalog_returns [cr_item_sk,cr_return_quantity,cr_returned_date_sk]
+                                            ReusedSubquery [d_date_sk] #1
                                     InputAdapter
-                                      ReusedExchange [i_item_sk,i_item_id] #2
+                                      ReusedExchange [i_item_sk,i_item_id] #5
                                 InputAdapter
-                                  ReusedExchange [d_date_sk] #3
+                                  ReusedExchange [d_date_sk] #2
         InputAdapter
           BroadcastExchange #8
             WholeStageCodegen (17)
@@ -85,7 +88,8 @@ TakeOrderedAndProject [item_id,sr_item_qty,sr_dev,cr_item_qty,cr_dev,wr_item_qty
                                   ColumnarToRow
                                     InputAdapter
                                       Scan parquet default.web_returns [wr_item_sk,wr_return_quantity,wr_returned_date_sk]
+                                        ReusedSubquery [d_date_sk] #1
                                 InputAdapter
-                                  ReusedExchange [i_item_sk,i_item_id] #2
+                                  ReusedExchange [i_item_sk,i_item_id] #5
                             InputAdapter
-                              ReusedExchange [d_date_sk] #3
+                              ReusedExchange [d_date_sk] #2