… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q58.sf100/explain.txt
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q58/explain.txt
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q83.sf100/explain.txt
… sql/core/src/test/resources/tpcds-plan-stability/approved-plans-v1_4/q83/explain.txt