@cloud-fan 2021-08-09
This is a tricky case, inferring more filters is generally good if it doesn't break DPP, so this is not a simple decision.

cc @maryannxue